Parse additional XML type information that can be used for encoding
